Introduce header spacing variables
