+++ interface.php
@@ -1,41 +1,10 @@
 <!-- Backend Variables -->
 <?php
-  function RGBtoHSL($color) {
-    $r = $color['red']; $g = $color['green']; $b = $color['blue'];
-    $r /= 255; $g /= 255; $b /= 255;
-    $max = max( $r, $g, $b );
-    $min = min( $r, $g, $b );
-    $h; $s; $l = ( $max + $min ) / 2;
-    $d = $max - $min;
-
-      if ($d == 0) { 
-        $h = $s = 0; 
-      } else {
-        $s = $d / ( 1 - abs( 2 * $l - 1 ) );
-        switch( $max ) {
-          case $r:
-            $h = 60 * fmod( ( ( $g - $b ) / $d ), 6 );
-            if ($b > $g) { $h += 360; }
-          break;
-
-          case $g:
-            $h = 60 * ( ( $b - $r ) / $d + 2 );
-          break;
-
-          case $b:
-            $h = 60 * ( ( $r - $g ) / $d + 4 );
-          break;
-        }
-      }
-
-    return array( 
-      'h' => round($h, 3) . 'deg', 
-      's' => round($s * 100, 3) . '%', 
-      'l' => round($l * 100, 3) . '%'
-    );
-  }
+  require_once 'includes/RGBtoHSL.php';
 
   $header_logo_height = get_field('fvt_header_logo_height', 'option');
+  $header_padding_top = get_field('fvt_header_padding_top', 'option');
+  $header_padding_bottom = get_field('fvt_header_padding_bottom', 'option');
 
   // Colors
   $bright_background_color = get_field('fvt_bright_background_color', 'option');
@@ -62,16 +31,10 @@
 ?>
 
 <style>
-  /* Imports */
-  /* Base Font */
   @import url('<?php echo $base_font_url; ?>');
-
-  <?php if ($display_font_url): ?>
-  /* Display Font */
-  @import url('<?php echo $display_font_url; ?>');
-  <?php endif; ?>
-  /* Variables */
-
+  <?php if ($display_font_url): ?>@import url('<?php echo $display_font_url; ?>');<?php endif; ?>
+  
+    /* Variables */
   :root {
     --shift: 10%;
 
@@ -122,6 +85,7 @@
     /* Utility Colors */
     --fvt-color-danger: hsl(357, 76%, 59%);
 
+
     /* Base Font */
     --fvt-base-font-family: <?php echo $base_font_family; ?>;
 
@@ -156,5 +120,7 @@
 
     /* Components */
     --fvt-header-logo-height: <?php echo $header_logo_height . 'px'; ?>;
+    --fvt-header-padding-top: <?php echo $header_padding_top . 'px'; ?>;
+    --fvt-header-padding-bottom: <?php echo $header_padding_bottom . 'px'; ?>;
   }
 </style>
