+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fouadvollmer/theme",
-  "version": "x.x.x",
+  "version": "dev",
   "description": "A theme for Wordpress by Fouad Vollmer",
   "main": "index.js",
   "scripts": {
