+++ style.css
@@ -4,6 +4,6 @@
  *  Author: Fouad Vollmer Werbeagentur
  *  Author URI: http://www.fouadvollmer.de
  *  Description: Individuelles Theme der Fouad Vollmer Werbeagentur
- *  Version: x.x.x
+ *  Version: dev
  *
  */
